# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API de Conversão de Moedas do Mercado de Wefin
  description: API para gerenciamento de taxas de câmbio e conversão entre Ouro Real e Tibar.
  version: 1.0.0

servers:
  - url: https://localhost:8080/api

security:
  - OAuth2:
      - read
      - write

paths:
  /oauth/token:
    post:
      summary: Autenticação do usuário
      description: Endpoint para autenticação OAuth2 e obtenção do token JWT.
      tags:
        - Autenticação
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  default: "password"
                username:
                  type: string
                password:
                  type: string
                client_id:
                  type: string
                client_secret:
                  type: string
              required:
                - grant_type
                - username
                - password
                - client_id
                - client_secret
      responses:
        '200':
          description: Token gerado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /coins:
    post:
      summary: Criar um nova moeda
      description: Cria um nova moeda na base de dados
      operationId: createCoin
      tags:
        - Moedas
      security:
        - OAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoinRequest'
      responses:
        '201':
          description: Moeda inserido com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoinResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado

    get:
      summary: Buscar as moedas
      description: Lista as moedas baseadas nos filtros informados
      operationId: searchCoins
      tags:
        - Moedas
      security:
        - OAuth2:
            - read
      responses:
        '200':
          description: Lista de moedas retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoinResponse'

  /coins/{uuid}:
    get:
      summary: Buscar uma moeda
      description: Busca uma moeda pelo ID informado
      operationId: findCoinById
      tags:
        - Moedas
      security:
        - OAuth2:
            - read
      parameters:
        - name: uuid
          in: path
          description: UUID da moeda
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Moeda encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoinResponse'
        '404':
          description: Recurso não encontrado

    put:
      summary: Atualizar uma moeda
      description: Atualiza todos os dados de uma moeda pelo ID
      operationId: updateCoin
      tags:
        - Moedas
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID da moeda
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoinRequest'
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoinResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

    delete:
      summary: Remover uma moeda
      description: Remove uma moeda pelo ID informado
      operationId: removeCoin
      tags:
        - Moedas
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID da moeda
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Remoção feita com sucesso
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

  /exchange-rates:
    post:
      summary: Criar uma nova taxa de câmbio
      description: Cria uma nova taxa de câmbio na base de dados
      operationId: createExchangeRate
      tags:
        - Taxas de câmbio
      security:
        - OAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRateRequest'
      responses:
        '201':
          description: Taxa de câmbio inserida com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRateResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '409':
          description: A taxa de câmbio informada já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Buscar as taxas de câmbio
      description: Lista as taxas de câmbio baseadas nos filtros informados
      operationId: searchExchangeRates
      tags:
        - Taxas de câmbio
      security:
        - OAuth2:
            - read
      responses:
        '200':
          description: Lista de taxas de câmbio retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeRateResponse'

  /exchange-rates/{uuid}:
    get:
      summary: Buscar uma taxa de câmbio
      description: Busca uma taxa de câmbio pelo ID informado
      operationId: findExchangeRateById
      tags:
        - Taxas de câmbio
      security:
        - OAuth2:
            - read
      parameters:
        - name: uuid
          in: path
          description: UUID da taxa de câmbio
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Taxa de câmbio encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRateResponse'
        '404':
          description: Recurso não encontrado

    put:
      summary: Atualizar uma taxa de câmbio
      description: Atualiza todos os dados de uma taxa de câmbio pelo ID
      operationId: updateExchangeRate
      tags:
        - Taxas de câmbio
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID da taxa de câmbio
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRateRequest'
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRateResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado
        '409':
          description: A taxa de câmbio informada já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Remover uma taxa de câmbio
      description: Remove uma taxa de câmbio pelo ID informado
      operationId: removeExchangeRate
      tags:
        - Taxas de câmbio
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID da taxa de câmbio
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Remoção feita com sucesso
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

  /products:
    post:
      summary: Criar um novo produto
      description: Cria um novo produto na base de dados
      operationId: createProduct
      tags:
        - Produtos
      security:
        - OAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Produto inserido com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado

    get:
      summary: Buscar os os produtos
      description: Lista os produtos baseado nos filtros informados
      operationId: searchProducts
      tags:
        - Produtos
      security:
        - OAuth2:
            - read
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

  /products/{uuid}:
    get:
      summary: Buscar um produto
      description: Busca um produto pelo ID informado
      operationId: findProductById
      tags:
        - Produtos
      security:
        - OAuth2:
            - read
      parameters:
        - name: uuid
          in: path
          description: UUID do produto
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Recurso não encontrado

    put:
      summary: Atualizar um produto
      description: Atualiza todos os dados de um produto pelo ID
      operationId: updateProduct
      tags:
        - Produtos
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID do produto
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

    delete:
      summary: Remover um produto
      description: Remove um produto pelo ID informado
      operationId: removeProduct
      tags:
        - Produtos
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID do produto
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Remoção feita com sucesso
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

  /kingdoms:
    post:
      summary: Criar um novo reino
      description: Cria um novo reino na base de dados
      operationId: createKingdom
      tags:
        - Reinos
      security:
        - OAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KingdomRequest'
      responses:
        '201':
          description: Reino inserido com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KingdomResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado

    get:
      summary: Buscar os reinos
      description: Lista os reinos baseado nos filtros informados
      operationId: searchKingdoms
      tags:
        - Reinos
      security:
        - OAuth2:
            - read
      responses:
        '200':
          description: Lista de reinos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KingdomResponse'

  /kingdoms/{uuid}:
    get:
      summary: Buscar um reino
      description: Busca um reino pelo ID informado
      operationId: findKingdomById
      tags:
        - Reinos
      security:
        - OAuth2:
            - read
      parameters:
        - name: uuid
          in: path
          description: UUID do reino
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reino encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KingdomResponse'
        '404':
          description: Recurso não encontrado

    put:
      summary: Atualizar um reino
      description: Atualiza todos os dados de um reino pelo ID
      operationId: updateKingdom
      tags:
        - Reinos
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID do reino
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KingdomRequest'
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KingdomResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

    delete:
      summary: Remover um reino
      description: Remove um reino pelo ID informado
      operationId: removeKingdom
      tags:
        - Reinos
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID do reino
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Remoção feita com sucesso
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

  /transactions:
    post:
      summary: Criar uma nova transação
      description: Cria uma nova transação na base de dados
      operationId: createTransaction
      tags:
        - Transações
      security:
        - OAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Transação inserida com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado

    get:
      summary: Buscar as transações
      description: Lista as transações baseada nos filtros informados
      operationId: searchTransactions
      tags:
        - Transações
      security:
        - OAuth2:
            - read
      responses:
        '200':
          description: Lista de transações retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'

  /transactions/{uuid}:
    get:
      summary: Buscar uma transação
      description: Busca uma transação pelo ID informado
      operationId: findTransactionById
      tags:
        - Transações
      security:
        - OAuth2:
            - read
      parameters:
        - name: uuid
          in: path
          description: UUID da transação
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transação encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Recurso não encontrado

    put:
      summary: Atualizar uma transação
      description: Atualiza todos os dados de uma transação pelo ID
      operationId: updateTransaction
      tags:
        - Transações
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID da transação
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

    delete:
      summary: Remover um transação
      description: Remove um transação pelo ID informado
      operationId: removeTransaction
      tags:
        - Transações
      security:
        - OAuth2:
            - write
      parameters:
        - name: uuid
          in: path
          description: UUID da transação
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Remoção feita com sucesso
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
        '404':
          description: Recurso não encontrado

components:
  schemas:
    CoinRequest:
      type: object
      properties:
        name:
          type: string
          description: Nome da moeda
      required:
        - name

    ExchangeRateRequest:
      type: object
      properties:
        idCoinTo:
          type: string
          format: uuid
          description: ID da moeda de destino
        idCoinFrom:
          type: string
          format: uuid
          description: ID da moeda de origem
        vlRate:
          type: number
          minimum: 0.01
          description: Uma moeda de origem valem X moedas de destino
      required:
        - idCoinTo
        - idCoinFrom
        - vlRate

    KingdomRequest:
      type: object
      properties:
        name:
          type: string
          description: Nome do reino
      required:
        - name

    ProductRequest:
      type: object
      properties:
        name:
          type: string
          description: Nome do produto
        vlUnit:
          type: number
          minimum: 0.01
          description: Valor unitário na moeda especificada
        idKingdom:
          type: string
          format: uuid
          description: ID do reino do produto
        idCoinBase:
          type: string
          format: uuid
          description: ID da moeda base do produto
      required:
        - name
        - vlUnit
        - idKingdom
        - idCoinBase

    TransactionRequest:
      type: object
      properties:
        quantity:
          type: number
          minimum: 1
          description: Quantidade de produtos comprados
        idProduct:
          type: string
          format: uuid
          description: ID do produto comprado
        idCoinUsed:
          type: string
          format: uuid
          description: ID da moeda usada no pagamento
      required:
        - qtBuyed
        - idProduct
        - idCoinUsed

    CoinResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID da moeda
        name:
          type: string
          description: Nome da moeda

    ExchangeRateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID da taxa de câmbio
        vlRate:
          type: number
          minimum: 0.01
          description: Uma moeda de origem valem X moedas de destino
        coinTo:
          $ref: '#/components/schemas/CoinResponse'
        coinFrom:
          $ref: '#/components/schemas/CoinResponse'

    KingdomResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID do reino
        name:
          type: string
          description: Nome do reino

    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID do produto
        name:
          type: string
          description: Nome do produto
        vlUnit:
          type: number
          minimum: 0.01
        kingdom:
          $ref: '#/components/schemas/KingdomResponse'
        coinBase:
          $ref: '#/components/schemas/CoinResponse'

    TransactionResponse:
      type: object
      properties:
        vlUnit:
          type: number
          minimum: 0.01
          description: Valor do produto no momento da transação
        vlRate:
          type: number
          minimum: 0.01
          description: Valor de câmbio utilizado na transação
        quantity:
          type: number
          minimum: 1
          description: Quantidade de produtos comprados
        vlConverted:
          type: number
          minimum: 0.01
          description: Valor pago após a conversão
        product:
          $ref: '#/components/schemas/ProductResponse'
        coinUsed:
          $ref: '#/components/schemas/CoinResponse'
        coinRequested:
          $ref: '#/components/schemas/CoinResponse'

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        fields:
          type: array
          items:
            type: string

  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/your-realm/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/your-realm/protocol/openid-connect/token
          scopes:
            read: Permissão para consultas (GET)
            write: Permissão para operações de escrita (POST/PUT/DELETE)